update features using latest dcontainer (#349)
* update features using latest dcontainer

* refactor: added more docs

+++ dev/generate_all_feature_definitions.py
@@ -1,5 +1,5 @@
 import os
-from dcontainer.cli.generate.generate_feature import generate
+from dcontainer.devcontainer.feature_generation.oci_feature_generator import OCIFeatureGenerator
 from pathlib import Path
 
 import typer
@@ -7,27 +7,33 @@ import shutil
 
 
 def generate_all_Feature_definitions(
-    feature_definitions_dir: str, output_dir: str
+    feature_definitions_dir: str, output_dir: str ,remove_old_content: bool = True,
 ) -> None:
-    
 
     for feature_name in os.listdir(feature_definitions_dir):
-        if os.path.exists(os.path.join(
+        if remove_old_content and os.path.exists(os.path.join(
                     "test", feature_name
                 ) ):
             shutil.rmtree(os.path.join(
                         "test", feature_name
                     )
                     )
+        if remove_old_content and os.path.exists(os.path.join(
+                    "src", feature_name
+                ) ):
+            shutil.rmtree(os.path.join(
+                        "src", feature_name
+                    )
+                    )
 
-        generate(
+        OCIFeatureGenerator.generate(
             Path(
                 os.path.join(
                     feature_definitions_dir, feature_name, "feature-definition.json"
                 )
             ),
             output_dir=Path(output_dir),
-            release_version="v0.2.3"
+            nanolayer_version="v0.4.0"
         )
